Fixed more warnings for GCC.

+++ code/icarus/BlockStream.cpp
@@ -582,7 +582,7 @@ int CBlockStream::Open( char *buffer, long size )
 
 	m_stream = buffer;
 
-	for ( int i = 0; i < sizeof( id_header ); i++ )
+	for ( size_t i = 0; i < sizeof( id_header ); i++ )
 	{
 		id_header[i] = *(m_stream + m_streamPos++);
 	}
